+++ src/genui/ui_widgets/editor_autocomplete.py
@@ -1,11 +1,13 @@
 from importlib.resources import open_text
 
-from PyQt6.QtWidgets import QCompleter, QTextEdit, QAbstractItemView
+from PyQt6.QtWidgets import QCompleter, QTextEdit, QAbstractItemView, QWidget
 from PyQt6.QtCore import Qt, QStringListModel, QRegularExpression
 from PyQt6.QtGui import QTextCursor, QPalette, QColor, QKeyEvent, QSyntaxHighlighter, QTextCharFormat, QFont
 
-from ..utils import Timer, BACKGROUND_COLOR_HEX
+from ..utils import BACKGROUND_COLOR_HEX
+from ..common.trace import Timer
 from ..settings import settings
+from .window_mixins.propagate_events import PropagateEventsMixin
 
 
 @Timer("Autocomplete words loader")
@@ -23,7 +25,7 @@ def load_words() -> list[str]:
     
     
 class PromptHighlighter(QSyntaxHighlighter):
-    def __init__(self, parent=None):
+    def __init__(self, parent: QWidget | None = None):
         super().__init__(parent)
         self.highlighting_rules = []
 
@@ -36,40 +38,42 @@ class PromptHighlighter(QSyntaxHighlighter):
         # case: `(picking (apricots)1.3)1.1, (apricots)1.1`
         self.add_rule(r"\([^,]+\)\d\.\d\b", None, settings.prompt_editor.compel_font_weight)
         # TODO: case: `off-topic,`
-        # TODO: case: `(koseki bijou), (ixy)0.7, (kanzarin)0.85, healthyman+, (quasarcake)0.5, (jonpei)0.9, (jima)1.1, realistic, (hi res)1.2, hololive, hololive english, (dongtan dress)1.3, (chest jewel)+`
+        # TODO: case: 
+            # `(koseki bijou), (ixy)0.7, (kanzarin)0.85, healthyman+, (quasarcake)0.5, (jonpei)0.9, 
+            # (jima)1.1, realistic, (hi res)1.2, hololive, hololive english, (dongtan dress)1.3, (chest jewel)+`
         
-    def add_rule(self, pattern, color: Qt.GlobalColor | None, weight: int | None):
+    def add_rule(self, pattern: str, color: Qt.GlobalColor | None, weight: int | None):
         regex = QRegularExpression(pattern)
-        format = QTextCharFormat()
+        format_rule = QTextCharFormat()
         
         if weight is not None:
-            format.setFontWeight(weight)
+            format_rule.setFontWeight(weight)
         if color:
-            format.setForeground(QColor(color))
+            format_rule.setForeground(QColor(color))
             
-        self.highlighting_rules.append((regex, format))
+        self.highlighting_rules.append((regex, format_rule))
 
-    def highlightBlock(self, text):
+    def highlightBlock(self, text: str):
         """Apply highlighting rules to the current block of text."""
-        for regex, format in self.highlighting_rules:
+        for regex, format_rule in self.highlighting_rules:
             match_iterator = regex.globalMatch(text)
             while match_iterator.hasNext():
                 match = match_iterator.next()
-                self.setFormat(match.capturedStart(), match.capturedLength(), format)
+                self.setFormat(match.capturedStart(), match.capturedLength(), format_rule)
     
     
 class WordsCompleter(QCompleter):
     words = load_words()
     completer_model = QStringListModel(words)
     
-    def __init__(self, parent=None):
+    def __init__(self, parent: QWidget | None = None):
         super().__init__(self.completer_model, parent)
         
         self.setCaseSensitivity(Qt.CaseSensitivity.CaseInsensitive)
         self.setFilterMode(Qt.MatchFlag.MatchContains)
 
 
-class AutoCompleteTextEdit(QTextEdit):
+class AutoCompleteTextEdit(QTextEdit, PropagateEventsMixin):
     min_word_length = 2
     
     def __init__(self, *args, **kwargs):
@@ -108,7 +112,7 @@ class AutoCompleteTextEdit(QTextEdit):
             Actual: "selected_word<start selection><cursor_position>,<end selection>"
         """
         word_under_cursor = cursor.selectedText()
-        is_compel_operator = word_under_cursor.startswith("+") or word_under_cursor.startswith("-")
+        is_compel_operator = word_under_cursor.startswith(("+", "-"))
         is_comma = word_under_cursor == ","
         if is_comma or is_compel_operator:
             cursor.movePosition(QTextCursor.MoveOperation.Left, QTextCursor.MoveMode.MoveAnchor, 2)
@@ -169,9 +173,8 @@ class AutoCompleteTextEdit(QTextEdit):
         self.completer.complete(cr) # Show popup
         
     def keyPressEvent(self, e: QKeyEvent):
-        if self.completer.popup().isVisible():
-            if e.key() == Qt.Key.Key_Return:
-                e.ignore()
-                return
+        if self.completer.popup().isVisible() and e.key() == Qt.Key.Key_Return:
+            e.ignore()
+            return
                 
         super().keyPressEvent(e)